Masmapas (#105)
* agrego configuracion docker

* hago mas abstracta la creacion de mapas

* masmapas

* Modifico Mapa y agrego clase ImagenMapa para agregar una imagen en el mapa!

* Agrego cartel de tunel y un tramo mas

* aumento la velocidad

* bastantes mas mapas agregados

* mapa1 terminado

* Empiezo mapa1 y mapa2

* arreglo mapa1

* mejoro consola

* agrego tapasui

* Hasta Sui Generis!

* mas mapas

* masfotoscharly

* agrego imagen partitura

* agrego hace4anios

* agrego mas velocidad cuando va caminando

* Creo funcion Hacer Mapa Historia Argentina

* Agrego fuente FriedNuget

* Agrego acelerador y divisor de velocidad

* Separo las funciones que crean los mapas en distintos archivos

* pinches

* rehagomapas

* Agregar dibujos de obstaculos mortales

---------
